quizzes now have hard deadline if due_at is set
due_at used to mean "the time after which you could no
longer start taking a quiz" but that's not really what
"due at" means.  We changed it instead to now mean what
it implies, "the time at which all quiz attempts must
be completed."

So if you start taking a 4-hour quiz one hour before
it's due, you'll only get one hour to take it.

The other part of this is fixing the timer to support
switching from counting up to counting down.  When an
assignment is untimed but due, say, two months from now,
we should count up on the timer.  When we get close to
the due date, though, this should change to a countDOWN
instead.  Right now this switch is set to happen at
12 hours from the due date, and will flash a notice
saying "12 hours left".

fixes #3727

Change-Id: I1eec558de61e2500da88f4127b7e691e7aa5c205
Reviewed-on: https://gerrit.instructure.com/2182
Reviewed-by: Brian Whitmer <brian@instructure.com>
Tested-by: Hudson <hudson@instructure.com>