fix outcome details to show correct numbers
On the outcomes page if you clicked "show details" for an
outcome, it would always say "0 artifacts" even if there
were artifacts.  The problem is it wasn't actually looking
for artifacts, it was looking for something else.  Also
fixed the pageless code on the outcome details page to
correctly do an infinite scroll.

fixes #3823

Change-Id: I6e0fd06ea70e5915ee5cdef23d4851ecd6a285fc
Reviewed-on: https://gerrit.instructure.com/2318
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>