get rid of 'show all' link on right side, add 'less' link
If there are more than 5 To Do items, it says "show more...".
If you clicked the link it would show up to 15 items.  If there
were more than that it would then show a "show all" link that
did nothing.  This link has been replaced with a "less..." link
that will shrink the list back down to 5 items.

fixes #4025

Change-Id: Ic9ef88dc62ccb0174ca402a8c8d102bd6967aa46
Reviewed-on: https://gerrit.instructure.com/2636
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>