don't send messages to the stream just because it was read by somebody
We were triggering "send_to_stream" even when just marking a ContextMessage as
read by somebody. That would trigger the message to pop to the top of all the
recipients' streams, which is confusing. This explains some, but probably not
all, of the odd stream item ordering that's been reported.

refs #4032

Change-Id: I43d9ae27aca02ccaf4d8939872360f4afedb4f1a
Reviewed-on: https://gerrit.instructure.com/2650
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>