deleting a user from an account sometimes deletes them from everything
If a user's logins are all part of one account and you enroll them in
a course in another account, then go to their page in that account and
you'll see a link to delete them from the root account. Deleting them
from that root account will delete them from everywhere, which is not
the intended behavior. The logic was only checking logins, not
enrollments, and also was just checking the number of root accounts, not
if there were any *different* root accounts.

fixes #4108

Change-Id: Id7584c8766420b3ae8bb557444785513dca201b9
Reviewed-on: https://gerrit.instructure.com/2819
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: JT Olds <jt@instructure.com>