Use html5 to show videos by default & show subtitles
using a library called mediaElement.js, try to show
all kaltura video and audio using html5 and fall back
on flash if needed.  fixes #3969

test plan:
in all the different browsers (including mobile safari/chrome)
try recording webcam videos (as homework submission, as content in a wiki page, etc)
view those same pages
view recorded submission in speedgrader

AND...
Allow uploading subtitle tracks to videos

when you upload a video, you will now see a [cc]
box in the player for that video if you are
the uploader of that video and in a modern browser.
it will have a link to upload a caption track in
SRT or WebVTT format in any language you specify.
once you attach a caption track to any of your
videos, other people that view it will be able
to choose that track while playing it.

helpful for people that are deaf or that speak
another language!

this change also adds google analytics tracking to
videos so we know when a given media_id was
played, paused, and ended.

test plan:
* record a video, go to play the video
* (only) you should see a [cc] button (make sure others don't)
* click the "upload subtitles" link, a dialog should appear
  that lets you choose a language and file to upload.  here's
  one you can test with:
  http://mediaelementjs.com/media/NT113_u008_v005_transcript.srt
* refresh, you (and anyone else that can see video)
  should now have the option to select that track
  for subtitles. (if using in HTML5 player)
* (only) you should see a delete 'x' on it to delete
  it, when you click it it should go away. reload
  to make sure it is not there.

Change-Id: I8ed8f14167e68c750815ec6496a9f992b744aa56
Reviewed-on: https://gerrit.instructure.com/2963
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>