use the same wording for 'my dashboard' in crumbs
the canvas logo in the top left as well as the home
icon in the breadcrubs both link to the dashboard.
no text is shown for either, just the image. but
it is confusing for people with screen readers since
they said different things but linked to the same
page.

Change-Id: Id75916b812fdd1ae5ab88528fdfcbc187bc120f8
Reviewed-on: https://gerrit.instructure.com/4919
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>