hide soon-to-be-dropped columns from activerecord, closes #5758
Whenever we create a migration that drops or renames a column, in that
same commit we need to add that column to the list in this initializer.

Then, when we deploy and restart, AR will already see the column as
removed, and we can safely run the drop migration after deploy without
any window where users will get page errors due to stale column
information.

Change-Id: I1c95fd146ac8046af579256cccf95b42f01442b7
Reviewed-on: https://gerrit.instructure.com/5823
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>