Fixes #5793 - Show question details issues
Old behavior:

The checkbox "Show question details" was displayed
all of the time, even w/ no questions on a new
quiz.

New Behavior:

The checkbox is hidden unless there are questions
that have anything to show (all but essay and
text).

It updates while you're editing, so if you delete
all of the questions, it hides, etc.

Change-Id: Ibb7bbfbadd1dcbdef0fc6c2807b1642f7454a789
Reviewed-on: https://gerrit.instructure.com/6265
Tested-by: Hudson <hudson@instructure.com>
Tested-by: Selenium <selenium@instructure.com>
Reviewed-by: JT Olds <jt@instructure.com>