drag-drop fix for new jquery and missing prop
there were two problems.

first, the jquery .bind() method supported the (type, listener,
useCapture) signature from addEventListener in older jQuery, but it was
undocumented. newer jQuery breaks that signature. take out the false
value intended for (but not used as) useCapture (that was the default
anyways) and it works.

second, the values for the keys in the params object for
$.toMultipartForm aren't always defined. for example, in
uploadAjaxFiles, the queued ajax upload doesn't always have the
duplicate_handling property set and so the params would include
'duplicate_handling': undefined. don't blow up in that case.

fixes #5663

Change-Id: Ib014eca887c5f3f935cc1d521e2184593d7e1bb8
Reviewed-on: https://gerrit.instructure.com/6273
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>