don't rename external tools when the tag is renamed
Other asset types (like assignments, wiki pages and files)
have a required linking between the name of the asset and
the name of the tag. This is for consistency's sake.
However, it doesn't make sense to enforce this kind of
policy for external tools, since tools can be configured
at the account level, and instructors shouldn't have the
ability to rename tools somewhere up the chain. This
change disconnects external tools from tags linking to
those external tools.

fixed #6170

testplan:
- configure an external tool
- add an external tool to a course module
- rename the link to something other than the tool's name
- check to make sure the tool's name didn't also change

- configure an external tool
- add an external tool to a course module
- rename the tool
- check to make sure the link's name didn't also change

- add an assignment to a course module
- rename the link in the module
- check to make sure the assignment's name also changed

- add an assignment to a course module
- rename the assignment
- check to make sure the module link's name also changed

Change-Id: I2291b4ca0ee2c83da1dd9a81bc46eb05d1925982
Reviewed-on: https://gerrit.instructure.com/6641
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: JT Olds <jt@instructure.com>