add sis id to add user dialog; fixes #6279
It only shows up when the admin adding the user has :manage_sis permissions,
and when the root account has sis imports enabled.

Test Plan:

 * Create new account
 * Go to Users page, verify that the Add User dialog doesn't include SIS ID
 * Enable sis imports on the account
 * Go to Users page, verify that the Add User dialog DOES include SIS ID
 * Remove the "manage SIS" privilege from your admin
 * Go to Users page, verify that the Add User dialog doesn't include SIS ID

The view spec covers these cases, and the controller spec covers testing that
if a user doesn't have :manage_sis, the sis id is ignored.

Change-Id: If8ec194772ecca99eaffb1195f117f87412041ae
Reviewed-on: https://gerrit.instructure.com/6845
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Whitmer <brian@instructure.com>