help dialog overhaul fixes: #5784
does a few things: consolidate the help
and support links at the bottom of the page
into 1 and calls them "help"

allows a school to set their own custom
help links to show in the help dialog so
they can link to their campus LMS support
in addition to our help options.

also cleans up the flow for teacher feedback and
submitting a ticket.

test plan:

* go to the account settings page of a root account
* you should see an option there to set up "custom
  help links", create a couple
* click the help link in the footer and the top left,
  a dialog should appear with the old options as well
  as the ones you set up.
* as a student, click the help link, click the teacher
  feedback option, fill out the form, the teacher AND TA's
  in that class should have got a canvasation.
* click the file a ticket option, fill out the form,
  zendesk should have a new issue with the
* and of course the specs in helpDialogSpec.coffee and
  help_dialog_spec.rb should pass

Change-Id: If2d841c00b28ece55eb118085ee8728b0de96c08
Reviewed-on: https://gerrit.instructure.com/7153
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>