translate links on imported assessment questions
Previously, assessment questions created via imports did not get their links
translated, since they are created with straight SQL. For now we will detect
when a question possibly has a link inside, and if it does, load the AQ model
and run translate_links on it.

This also includes a fix to translate_links to find them inside arrays too
(like in the answers array).

testplan:
 * import a QTI package with images in the questions and answer feedback
 * verify that the new questions in your question banks have links pointing to
   /assessment_questions/ and not /courses/ for the images
 * the existing translate_links specs cover the translate_links refactor

Change-Id: Idf36e65ef6a9887369feb1b2cdcbb2db1f2f70f3
Reviewed-on: https://gerrit.instructure.com/7136
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>