fixes qti file importing for packages without full path references
Some QTI packages don't correctly reference needed files. For
example, a package may have images 5 sub-directories down, but
a question item's reference won't start from the root but from
some sub-directory down the tree. This commit will check the
suffix of all the file paths when converting the html of a
question so that the correct file can be found.

The same packages that generally have this problem use the same
naming scheme for each package. So an image in two different
packages will have the same file path in the export. This commit
puts each import into its own directory so that different files
with the same path won't overwrite each other

Test Plan:
 * Import the plain_qti.zip fixture into a course
 * Make sure there are 4 questions on the imported quiz with images
 * Make sure the images are in the directory "Quiz Files/plain_qti_##/"

closes #5178 #6206

Change-Id: Ia48bd36b3ac64c74fe58c769942b256d4fae6552
Reviewed-on: https://gerrit.instructure.com/7187
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>