make wait_for_ajax and wait_for_animations not poll any more
Now we listen for events to know when ajax and animations are over, which may
speed up some checks by a couple hundred milliseconds. The caller can also
optionally specify an interval (in ms) to wait for the ajax/animation to
start. This is useful for tests that have ajax calls that don't start
synchronously.

test plan: run the selenium specs.

Change-Id: Ie9ce3799108db69ae65b0ae0f4a99420e5cdc711
Reviewed-on: https://gerrit.instructure.com/7256
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Tested-by: Hudson <hudson@instructure.com>