don't prepend csrf json protection to oauth2 post response
Even if the poster still has a session here, the response should never
be prepended since this is part of the API.

Also, the session was persisting too long, which is how we discovered
this problem. The session should be getting reset even if the native app
doesn't actually do the GET for the page that includes the ?code= param.

test plan: Go through the oauth2 native application flow, the JSON
response when you POST to trade the code for the access_token should not
have "while(1);" prepended to it.

Change-Id: I9db4ec2d2fd9dbfdb489081e4aeb36e8390b9bd6
Reviewed-on: https://gerrit.instructure.com/7260
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>