move the x-ua-compatible meta tag to an actual http header
It wasn't taking effect, because we're checking the IE version in order
to add classes to the <html> tag -- so by the time we hit the meta tag,
it's too late to switch. That's our theory, anyway.

Some IE widgets embedded in applications are displaying Canvas using
IE7 mode, which causes loads of issues. This tells them to use the
newest installed IE mode.

test plan: visit Canvas in IE 8 or 9 with compatibility mode enabled,
you should still not see the warning about upgrading because you are on
IE7. Also try the LDB, if available.

Change-Id: I03fd72c108e17f60f68b63445f864d135c29b32e
Reviewed-on: https://gerrit.instructure.com/7325
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>