don't queue job for attachments that won't be scribd anyway
previously we were always queueing the job, and then just changing the
workflow state and doing nothing else for attachments that aren't
scribdable. this creates a ton of needless jobs, especially on big
content imports.

now we just skip the background job completely when the attachment isn't
scribdable.

this will still generate the update_associated_content_tag jobs for all
these attachments.

test plan: upload a scribadble file, such as a pdf, and see that it is
submitted to scribd in a delayed job. upload a non-scribdable file, such
as a png or xml file, and see that no job is even created.

fixes #6619

Change-Id: I20f6bfda69cf9b12a7462b22f0674d3cac5d2d88
Reviewed-on: https://gerrit.instructure.com/7474
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>